backend/requirements.txt
# =====================================
# 🧠 CORE FRAMEWORK
# =====================================
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0

# OR if using FastAPI instead of Flask:
# FastAPI==0.115.0
# Uvicorn==0.30.1

# =====================================
# 🗄️ DATABASE & CONFIG
# =====================================
pymongo==4.15.3
python-dotenv==1.0.1
dnspython==2.6.1

# =====================================
# 🔐 SECURITY & AUTHENTICATION
# =====================================
bcrypt==4.1.2
PyJWT==2.8.0

# =====================================
# 🤖 AI / MACHINE LEARNING MODULE
# =====================================
numpy==2.1.2                 # 2.x wheels officially support Python 3.13
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2

# =====================================
# 🧪 TESTING
# =====================================
pytest==8.3.3
requests==2.32.3

# =====================================
# 🧰 UTILITIES / MISC
# =====================================
Werkzeug==3.0.3
itsdangerous==2.2.0
click==8.1.7
gunicorn==22.0.0

# =====================================
# 💬 REAL-TIME / OPTIONAL MODULES
# =====================================
flask-socketio==5.3.6
eventlet==0.35.2

python-multipart==0.0.6
# =====================================
